Simplify types in loop-unbounded-array-complex

+++ basic/loop-unbounded-array-complex/example.c
@@ -1,9 +1,9 @@
-unsigned long example(unsigned long n) {
-  unsigned long data[4] = { 0, 1, 2, 3 };
-  for (unsigned long i = 0; i < n; i++) {
-    unsigned long i2 = (i + 2) % 4;
-    unsigned long *a = &data[i];
-    unsigned long *b = &data[i2];
+int example(int n) {
+  int data[4] = { 0, 1, 2, 3 };
+  for (int i = 0; i < n; i++) {
+    int i2 = (i + 2) % 4;
+    int *a = &data[i];
+    int *b = &data[i2];
     *b += n;
   }
   return data[n % 4];
